Print a rough estimation of how long the build will still last.
Every time the build status is printed, perform a simulation of the rest of the
build so we can add an estimation to the build status. For this simulation,
historic build times from the Ninja log are used. If a build is started without
any Ninja log present, an estimation cannot yet be made.

To do:
* Update the BuildLog with our runtimes for every finished job, and write the
  log to .ninja_log at exit. This will allow immediate showing of estimations
  even without a pre-existing ninja log, and will make future estimations more
  accurate.
* Move this code into a separate function that receives queue and worker state,
  and returns an Option<Duration>.
* Since this code may take a long time to run in larger projects, perhaps we
  shouldn't run it for every print of the build status, but allow it to
  recompute less frequently and just print the estimation minus elapsed time
  since calculation.
* Add a pretty progress bar.