Improve display of the current progress.
This patch adds the -P / --progress option to ninj, which sets the type of
progress display. There are currently four options: no progress display (none),
display in text (text), display a progress bar in ASCII (ascii) or use fancy
Unicode characters (highres).

The term_size dependency is added for establishing the width of the terminal.

A bug is fixed causing the final time estimation to be unknown, caused by all
workers being Done.